Create index.php
Installed php and copied the path in environment variables. This is as below 
C:\xampp\php;C\System\.........
Does not run in browser and 'php not found error'. Don't know what is wrong.
Default localhost browser not working.

Beginner in PHP. Please help, how to run this file and setting the environment variables correctly. Thanks in advance.